# Conflicts:
#	.travis.yml

+++ src/main/java/com/mmnaseri/utils/spring/data/query/impl/DefaultDataFunctionRegistry.java
@@ -1,5 +1,6 @@
 package com.mmnaseri.utils.spring.data.query.impl;
 
+import com.mmnaseri.utils.spring.data.error.DuplicateFunctionException;
 import com.mmnaseri.utils.spring.data.error.FunctionNotFoundException;
 import com.mmnaseri.utils.spring.data.query.DataFunction;
 import com.mmnaseri.utils.spring.data.query.DataFunctionRegistry;
@@ -9,21 +10,34 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
- * @author Mohammad Milad Naseri (m.m.naseri@gmail.com)
+ * This class provides support for registering data functions. Also, it comes with the option to register
+ * the default functions out-of-the-box.
+ *
+ * @author Milad Naseri (mmnaseri@programmer.net)
  * @since 1.0 (9/29/15)
  */
+@SuppressWarnings("WeakerAccess")
 public class DefaultDataFunctionRegistry implements DataFunctionRegistry {
 
     private final Map<String, DataFunction<?>> functions;
 
     public DefaultDataFunctionRegistry() {
-        functions = new ConcurrentHashMap<String, DataFunction<?>>();
-        register("count", new CountDataFunction());
-        register("delete", new DeleteDataFunction());
+        this(true);
+    }
+
+    public DefaultDataFunctionRegistry(boolean registerDefaults) {
+        functions = new ConcurrentHashMap<>();
+        if (registerDefaults) {
+            register("count", new CountDataFunction());
+            register("delete", new DeleteDataFunction());
+        }
     }
 
     @Override
     public void register(String name, DataFunction<?> function) {
+        if (functions.containsKey(name)) {
+            throw new DuplicateFunctionException(name);
+        }
         functions.put(name, function);
     }
